+++ client/src/sfml/SoundSFML.cpp
@@ -33,7 +33,7 @@ void SoundSFML::stop()
     _sound.stop();
 }
 
-void SoundSFML::setVolume(float volume)
+void SoundSFML::setVolume(double volume)
 {
     ASound::setVolume(volume);
     _sound.setVolume(volume);
