+++ engine/include/graphical/IAudio.hpp
@@ -21,7 +21,7 @@ namespace Engine {
         virtual void play() = 0;
         virtual void pause() = 0;
         virtual void stop() = 0;
-        virtual void setVolume(float volume) = 0;
+        virtual void setVolume(double volume) = 0;
     };
 
 }
