+++ engine/include/sceneManagement/AScene.hpp
@@ -31,7 +31,7 @@ namespace Engine {
         virtual void onFocus() {};
         virtual void onExit() {};
 
-        void updateDeltatime(float delta);
+        void updateDeltatime(double delta);
         [[nodiscard]] int getId() const;
         void spawnEntity(std::shared_ptr<Entity> &entity);
         void despawnEntity(std::shared_ptr<Entity> &entity);
