+++ engine/include/tools/Timer.hpp
@@ -17,8 +17,8 @@ namespace Engine {
         Timer();
         ~Timer() = default;
 
-        [[nodiscard]] static bool hasElapsed(std::chrono::high_resolution_clock::time_point time, float seconds);
-        float deltatime(float multiplier = 0.01);
+        [[nodiscard]] static bool hasElapsed(std::chrono::high_resolution_clock::time_point time, double seconds);
+        double deltatime(double multiplier = 0.01);
 
         const std::chrono::high_resolution_clock::time_point &getLastPoint() const;
     };
